% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cablecuttr.R
\docType{package}
\name{cablecuttr}
\alias{cablecuttr}
\alias{cablecuttr-package}
\title{cablecuttr: A CanIStream.It API Wrapper}
\description{
An R wrapper for the \href{http://www.canistream.it/}{CanIStream.It}
  API for searching across the most popular streaming, rental, and purchase
  services to find where a movie is available.
}
\section{Querying functions}{

\describe{
  \item{\code{\link{find_movie}}}{Searches CanIStream.It for the movie}
  \item{\code{\link{can_i_stream}}}{Finds which services/stores the movie is available on}
}
}

